requirements.txt
openai
groq
anthropic
python-dotenv
tiktoken
exa_py
pypdf